fixed problem with looping through everything + added string seperator for answers
when you looped through all problems and fixed them all, you got stuck in the loop, because continue_labeling was only set to False, if you chose "Done" and not if you chose "Next". You got stuck, because there was nothing left to label. Added a new function for DataFrame, which counts all unlabelled lines. if there are 0, continue_labelling is set to False.

Small extra change: Adding the string seperator to ET.
if there is a string_seperator in an answer, the answer is seperated in two parts and there are searched through seperately. the results are still together for the column they were in, but "not so, sad" would only give us the result "sad" and not "not so sad"